1.修复webview不在同一线程的报错 2.修复activiity结束dialog未关闭报错 3.修复控件为null 4.修复bookutil 文件为空 5.修复glideImageloader file不为空

+++ app/src/main/java/com/heiheilianzai/app/utils/BookUtil.java
@@ -211,8 +211,10 @@ public class BookUtil {
             file.mkdir();
         } else {
             File[] files = file.listFiles();
-            for (int i = 0; i < files.length; i++) {
-                files[i].delete();
+            if (files != null && files.length > 0) {
+                for (int i = 0; i < files.length; i++) {
+                    files[i].delete();
+                }
             }
         }
     }
